refact(PriceCalculatorFooter): PriceCalculatorFooter --> PriceCalculatorFooterForm (#510)
